[RED] create stub and test for send_mass_html_mail
send_mass_html_mail is a function that will be used to send mass mail
with html as its content. This function will be build on top of django
mail module. This function is required as django's default mass mail
service does not allow for html content and only uses plain text. This
prevents us from including images and formatting the text payload.