[GREEN] fix test of one time code for assessor registration
A set of tests for assessor registration, involving one time code, has
been created earlier. However, two of the tests were problematic as it
attempts to produce a non registered code by replacing the last
character with the first character. Such approach may lead to
nondeterministic behavior as the last character may be equal to the
first one, hence resulting to the same code. This commit replaces it
with a new generated UUID that is guaranteed to be different.