[REFACTOR] move all exceptions under exceptions.py
Before this commit, all of the exceptions are scattered across the
different applications. However, this implementation may result to
duplication as we often not know whether an exception already exists.
Hence, grouping the exception will give a better understanding on what's
available and what's not, and make it easier to maintain.