PBI-06 Company Signup (#24)
* [RED] create test for email with same case functionality

In the previous implementation, there is a hidden bug as email with same
characters but different case can be used to register 2 different
accounts. This commit tests the validation and controller method that
serves the functionality for COMPANY.

* [GREEN] implement functionalities to prevent duplicated email

Store email in its lowercase form and perform validation to the
normalized version of email.