PBI-20 Add Event Participants (#33)
* [RED] create test for email with same case functionality

In the previous implementation, there is a hidden bug as email with same
characters but different case can be used to register 2 different
accounts. This commit tests the validation and controller method that
serves the functionality for COMPANY.

* [GREEN] implement functionalities to prevent duplicated email

Store email in its lowercase form and perform validation to the
normalized version of email.

* [RED] create stub and test for serve_get_user_info

* [GREEN] implement serve_get_user_info functionalities

The function will be used to aid protecting routes at the front end. It
will return the data of the authenticated user.

* [RED] create model for test flow, test and stub for utility functions

Create TestFlow and TestFlowTool models for TestFlow creation. The
serializers were also created using the default
serializer.ModelSerializer approach. As TestFlowTool will deals heavily
with AssessmentTool and such distinction between subclasses of
AssessmentTool is required, AssessmentTool now inherits from Polymorphic
model. This commit also creates stub and test for utility function
needed for test flow creation.

* [CHORES] add django-polymorphic to requirements.txt

* [GREEN] implement utility functions for creating test flow

* [RED] create test and stub for adding tool to test flow

* [GREEN] implement add_tool function of TestFlow

* [RED] create stub and test for test flow registration validator

* [GREEN] implement validate_test_flow_registration

* [RED] create test for convert_assessment_tool_id_to_assessment_tool

One of the request attribute is a list of assessment tool id and release
time. This function will be used to convert each of the id to its
respective assessment tool and the release time to be a time object.

* [GREEN] implement convert_assessment_tool_id_to_assessment_tool

* [RED] create tests for save_test_flow_to_database

* [GREEN] implement save_test_flow_to_database

* [RED] add end to end tests for test flow creation

* [GREEN] implement end-point to create test flow

Allow company or assessor to create test flow by submitting request
according to the structure specified.

* [REFACTOR] use InvalidTestFlowRegistration for test flow exception

Earlier, InvalidRequestException was used. To aid better error handling
and to ensure that error handling become more apparent, a more specific
class is introduced. To ensure backward compatibility, the exception
still inherits from InvalidRequestExceptio use
InvalidTestFlowRegistration for test flow exception

Earlier, InvalidRequestException was used. To aid better error handling
and to ensure that error handling become more apparent, a more specific
class is introduced. To ensure backward compatibility, the exception
still inherits from InvalidRequestException.

* [RED] create stub and test for assessment tool authorization

Previously, company can use any assessment tool registered in the
database. This create stubs for the necessary functions that prevents that
from happening by ensuring that only tools associated with the company
are retrieved in the previously 'get_tool_by_id' function.

* [GREEN] implement get_tool_of_company_from_id

* [RED] add test for get_time_from_date_time for None

* [GREEN] implement get_time_from_date_time_string to handle None

* [RED] create tests for adding start_working_time in TestFlow

start_working_time field indicates when will the assessee can first
start doing the activity. This commit create the necessary tests to
evaluate the to be made changes on services and model methods.

* [GREEN] implement functionalities to accomodate start_working_time

* [REFACTOR] change RestrictedAccessException code from 401 to 403

* [REFACTOR] refactor add_tool method to persist changes

* [RED] model and create tests for assessment event registration

* [GREEN] implement assessment event registration

* [REFACTOR] extract error message to prevent duplication

* [REFACTOR] remove unused import statemments

* [RED] create stub and models to allow adding participants to events

* [RED] add stubs for methods in AssessmentEvent model

* [RED] fix test logic for participant registration

* [GREEN] implement functionalities to allow adding participants to events