[REFACTOR] change authorization exception response to forbidden
Previously, if a user already login but failed to gain authorization to
a certain service, an AuthorizationException with status 401 will be
raised. However, this conflicts with the front end mechanism to refetch
the access and refresh token. This commit changes all previously
AuthorizationException to RestrictedAccessException and changes the
response code from UNAUTHORIZED to FORBIDDEN.