PBI-03 Submitting Assignments (#49)
* [RED] create test for assessor_dashboard

* [GREEN] create stub for assessor_dashboard

* [REFACTOR] explicitly define authorized HTTP methods

* [REFACTOR] remove unused imports

* [RED] create tests for fetching active assessees of an assessor

create tests for the endpoint to get active assessees associated with the assessor and methods in dashboard.py service

* [GREEN] implement the get an assessor's active assessees feature

* [REFACTOR] rename ActiveAssessmentToolTest

rename ActiveAssessmentToolTest to ActiveAssesseesTest in assessor module's tests.py to better describe what the class tests

* [RED] create test for endpoint to fetch assessment events of an assessor

create tests for fetch assessment endpoint and functions in services

* [GREEN] implement fetch assessment events for assessor feature

* [GREEN] fix test of assessor due to changes in status code convention

* [REFACTOR] change event does not exist error message for consistency

Previously, there are 2 different error messages for event that does not
exist. This commit standardize them into 1 for consistency.

* [RED] create test and stub for get_assessment_event_participation_by_assessee

Method is introduced to get event participation by assessee from an assessment
event.

* [GREEN] implement get_assessment_event_participation_by_assessee

* [RED] create AssignmentAttempt model, stub and test for get_assignment_attempt

Get assignment attempt is a method of event-participation model to
get attempt corresponding to an assignment. In addition, this commit
also add a saving mechanism to store the test flow attempt associated
with the AssessmentEventParticipation object.

* [GREEN] implement get_assignment_attempt of AssessmentEventParticipation

* [RED] create stub and test for create_assignment_attempt

The method is dedicated to create the attempt for an assessee for an
assignment he/she undergoes in an event.

* [GREEN] implement create_assignment_attempt

* [RED] create stub and test for upload_file_to_google_bucket

This function will interact with the google-cloud file upload services.
It needs to invoke several functions as defined in the Google's
documentation.

* [GREEN] implement upload_file_to_google_bucket

* [CHORES] update requirements.txt to include google modules

* [RED] create stub and tests for AssignmentAttempt methods

Create mutating methods: update_attempt_cloud_directory and
update_file_name.

* [CHORES] update gitflow to include google credentials env

* [CHORES] add google credentials to collectstatic workflow

GOOGLE_APPLICATION_CREDENTIAL_VALUES is required each time the
application is run, including collectstatic. Hence, it needs to be
present at Collect Static step in Github workflow.

* [GREEN] remove test and stub for AssignmentAttempt methods

This is done to ensure correctness of upload_file_to_google_bucket
method following the TDD approach. This commit is expected to pass the
tests for upload_file_to_google_bucket.

* [RED] re-crate stub and test for AssignmentAttempt methods

Previously, it was removed due to the commit being intervening the
RED-GREEN flow of upload_file_to_google_bucket. This commit adds a new
test to ensure correctness of changing the submitted time in
AssignmentAttempts.

* [GREEN] implement methods of AssignmentAttempt

* [RED] create 'facade' function save_assignment_attempt

The method will be used as the sole caller to save the assignment
attempt to the database and upload it to the Google Storage. This method
will call the corresponding functions.

* [GREEN] implement save_assignment_attempt of assessment_event_attempt

* [RED] create stub and tests for get_assessment_tool_from_assessment_id

The method will be used to retrieve assessment tool associated with an
event from the given ID.

* [GREEN] implement get_assessment_tool_from_assessment_id

* [RED] create stub and tests for validate_submission

The validate_submission function will be used to ensure correctness of
filename in the submitted file according to the set expectations in the
assignment creation.

* [GREEN] implement submission validator

* [RED] create end-to-end tests for assignment submission

* [GREEN] implement end-point to submit assignments

* [REFACTOR] refactor test by removing duplication and renaming shadowed variables

Co-authored-by: indiradr <indira.devi@ui.ac.id>