[RED] create stub and test for get_event_end_date_time
get_event_end_date_time will be used to get the datetime of an
assessment event using the UTC timezone. It is computed by adding a
buffer 10 minutes from the last assessment tool deadline in the
underlying test flow. This method will be used to help submission
validation in the front end, as well as to determine whether an
assessment event is active.