[RED] create stub, endpoint, and test to get assessment event data
This function will be used to return assessment event data, and most
importantly, the end datetime of an assessment event. To support this
point, the AssessmentEventSerializer has been modified to accomodate the
endtime datetimefield. With respect to the team's request on providing a
consistent ISO date, the start and end time data will be presented in
ISO format.