PBI-07 Company Assessor Login (#63)
* [RED] create tests and stub for get_assessor_from_email

The function will be used to return the corresponding assessor from the
given email.

* [GREEN] implement get_assessor_from_email

* [RED] create stub and test for get_company_from_email

The method returns the company object associated with the given email.

* [GREEN] implement get_company_from_email

* [RED] create stub and tests get get_assessor_or_company_from_email

The function will be used to retrieve either assessee or company that
matches the email. The test uses mock as the function depends on the
previous 2 functions.

* [GREEN] implement get_assessor_or_company_from_email

* [RED] create stub and tests for verify_password

verify_password function is used as a validator to check user's
password. It will raise exception when the password does not match and
does not return nor raise exception when the password matches.

* [GREEN] implement verify_password

* [RED] create test and stub for get_assessor_or_company_from_request_data

The method will be used to raise InvalidRequestException when no
assessor or company with matching email is found and return the user
object when the user is found.

* [GREEN] implement get_assessor_or_company_from_request_data

* [RED] create tests, endpoint and facade function to get login tokens

This commit creates the endpoint placeholder that will return the access and refresh token to a
valid request and create the end-to-end test.

* [GREEN] implement serve_login_assessor_company

* [REFACTOR] remove duplication in tests and change convention in models

Remove duplicated strings in tests.py and follow python class method
convention for the AssessmentEventSerializer get_end_time_iso and
get_start_time_iso by using self instead of obj as the first parameter.