[RED] create stub and tests for validate_assessor_responsibility
This commit creates the test and stub for validate_assessor_responsibility.
The function will be used to ensure that the requesting user is the
assessor that is responsible for the assessee that creates the attempt.
The function will raise RestrictedAccessException which later will
trigger a 403 response in the views functioThe function will raise
RestrictedAccessException which later will trigger a 403 response in the
views function.