[REFACTOR] refactor serve_get_submitted_assignment
serve_get_submitted_assignment previously defined the response within
the views function. However, a utility function to wrap the returned
file has been created and present in the utils module as it the function
is used in the assessor's end-point to retrieve assessee's submitted assignment.
This commit replaces the definition in the views with using the utils
function.