PBI-17 Grading Assessment (#70)
* [RED] create stub and tests for validate_grade_assessment_tool_request

This commit creates the tests for
validate_grade_assessment_tool_request.
The function will be part of the grading end-point and serves as the
validator to ensures that the request is valid.

* [GREEN] implement validate_grade_assessment_tool_request

* [RED] create stub and tests for validate_assessor_responsibility

This commit creates the test and stub for validate_assessor_responsibility.
The function will be used to ensure that the requesting user is the
assessor that is responsible for the assessee that creates the attempt.
The function will raise RestrictedAccessException which later will
trigger a 403 response in the views functioThe function will raise
RestrictedAccessException which later will trigger a 403 response in the
views function.

* [GREEN] implement validate_assessor_responsibility

* [CHORES] add note field and setters to ToolAttempt model

* [RED] create tests and stub for set_grade_and_note_of_tool_attempt

This commit create tests and stub for the
set_grade_and_note_of_tool_attempt function. The function will mutate
the assessment tool attempt according to the grade and note given in the
request data.

* [GREEN] implement set_grade_and_note_of_tool_attempt

* [RED] create end-to-end test for serve_grade_assessment_tool_attempts

This commit creates the end point for submitting grade and notes for an
assessment tool attempt created by assessee.

* [GREEN] implement end-point to submit grade and note for tool attempt

* [RED] create tests for serve_get_assignment_attempt_data

The serve_get_assignment_attempt_data function will be used as an
end-point handler to return the data of the assignment attempt made by
the assessee.

* [GREEN] implement serve_get_assignment_attempt_data

* [RED] create tests and stubs for end-point to download submitted attempt

The endpoint will be used by assessors to download the assessee's
attempts on the assignment used in the assessment event. There are no
service level tests created because most of the functions used has
already been implemented and tested.

* [GREEN] implement serve_get_assignment_attempt_file end-point

* [REFACTOR] refactor serve_get_submitted_assignment

serve_get_submitted_assignment previously defined the response within
the views function. However, a utility function to wrap the returned
file has been created and present in the utils module as it the function
is used in the assessor's end-point to retrieve assessee's submitted assignment.
This commit replaces the definition in the views with using the utils
function.

* [REFACTOR] introduce a decorator to convert exceptions to InvalidRequest

Some of the service functions explicitly catching
ObjectDoesNotExistException and converting them to
InvalidRequestException. This commit extracts those conversion process
to a separate function that is used as a decorator.