PBI-20: Modifying Assessment Event (#72)
* [RED] create stub and tests for validate_update_assessment_event

The validate_update_assessment_event will be used to ensure that the
assessment event update request is valid. Done by ensuring that the
start time is valid, name does not violate any constraint, and the test
flow id exists.

* [GREEN] implement validate_update_assessment_event

* [RED] create stub and tests for update_assessment_event_from_request_data

The function will be used to save the new data presented in the
request_data to the assessment event object. The function will use the
setter methods of the AssessmentEvent.

* [GREEN] implement update_assessment_event_from_request_data

* [RED] create stub and end-to-end test for updating assessment event

* [GREEN] implement serve_update_assessment_event

* [RED] create stub and tests for has_attempted_test_flow

The function will return true if the assessee has submitted any attempt
for the assessment tools used in the assessment event and false if
otherwise.

* [GREEN] implement has_attempted_test_flow of AssessmentEventParticipation

* [RED] create stub and tests for AssessmentEvent's has_been_attempted

The method will return true if a ToolAttempt has been submitted to any
of the assessment tools used in the assessment event and false if
otherwise.

* [GREEN] implement AssessmentEvent's has_been_attempted

* [RED] create stub and tests for start_time_has_passed

The method will be used to check whether an AssessmentEvent is
deletable. It will return true if the assessment event start time has
passed and false if otherwise.

* [GREEN] implement start_time_has_passed

* [RED] create stub and tests for is_deletable

The is_deletable method will be used in the assessment event deletable
mechanism. An assessment event is only deletable when the start time has
not passed and no attempt has been made for the event.

* [GREEN] implement AssessmentEvent's is_deletable

* [RED] create stub and tests for validate_delete_assessment_event_request

The function is required in the deletion of an assessment event. When
the assessment event is not deletable, the function should raise an
exception. Otherwise, it should not return or raise anything.

* [GREEN] implement validate_delete_assessment_event_request

* [RED] create stub and end-to-end tests for deleting assessment event

* [GREEN] implement end-point to delete AssessmentEvent