PBI-17: Interactive Quiz Grading (#76)
* [RED] create stub and test for assessor registration

* [RED] create stub and test for assessor registration views

* [RED] create stub and test for generate one time code

* [RED] fix expected message for test_register_assessor_with_invalid_one_time_code function

* [GREEN] implement assessor registration and one time code generation

* [CHORES] set initial migration dependencies

* [CHORES] insert phonenumbers dependency into requirements.txt

* [REFACTOR] remove unused variable in test_validate_user_assessor_registration_data_with_invalid_one_time_code

* [REFACTOR] remove unused response variables in users/test.py

* [RED] create tests for assessor google registration

create test for assessor gooogle registration endpoint in views.py, create tests for functions in google_login service for assessor

* [GREEN] implement assessor registration by google feature

* [REFACTOR] create constant USER_IS_ALREADY_REGISTERED in tests.py

define a constant as per SonarCloud's code smell warning to not duplicate string literals

* [CHORES] resolve merge conflict from merge with staging

* [RED] create tests for grading interactive quiz and get interactive quiz attempt data

* [REFACTOR] change interactive quiz grading test class name and uncomment attempt tests

* [GREEN] implement interactive quiz grading feature and get interactive quiz data feature

* [GREEN] remake migrations to accomodate new models

* [GREEN] remake migrations to fix dependency conflicts

* [GREEN] remake migrations to fix dependency conflicts

* [REFACTOR] reorder views function based on their functionality

This commit regroup the functions in the views.py to make it easier for
developers to search the function they need.

* [RED] create test and stub for get_assessee_user_with_google_matching_data

The function will be used to get the assessee object that is associated
to the google user data. This function will be used in the assessee login
mechanism.

* [GREEN] implement get_assessee_user_with_google_matching_data

* [RED] implement end-to-end test for assessee google login

* [GREEN] implement serve_google_login_callback_for_assessee

* [REFACTOR] remove duplication by extracting mock setup as independent function

* [RED] modify test to allow function to raise InvalidRequestException

Previously, the login service function throw a EmailNotFoundException
which converts to a response with Internal Server Error status. Thiis
commit modifies the test to allow InvalidRequestException to be raised
instead.

* [GREEN] modify get_assessee_user_with_google_matching_data

Modify get_assessee_user_with_google_matching_data to return
InvalidRequestException such that a BAD_REQUEST response will be raised
when the user with email is not found.

* [RED] create test and stub for get_assessor_user_with_google_matching_data

The function will be used to get the associated user from the given
google user data.

* [GREEN] implement get_assessor_user_with_google_matching_data

* [RED] create end-to-end test for assessor to obtain token through glogin

* [GREEN] implement serve_google_login_callback_for_assessor

* [REFACTOR] reduce duplicated code in tests.py in the assessment module

create constants for duplicated dictionary and update values for instances

* [REFACTOR] reduce code smells based on SonarCloud analysis

remove duplicated string literals and reduce cognitive complexity of functions

* [REFACTOR] use constants to define URL instead of class attrib in tests

* [REFACTOR] modify registration of assessee and assessor through google

Change the getter to get_[assessee|assessor]_with_google_matching_data
rather than the combined version due to the separate registration and
login end-point for assessee and assessors.

* [REFACTOR] change password parameter value
change password parameter when creating users so SonarCloud doesn't raise a security hotspot issue

* [REFACTOR] repurpose google register end-point for both login and register

The front end application only have 1 button for assessee to login using
google login. Hence, if the user has been registered, a login operation
will be conducted. On the other hand, when a new user data is obtained,
a register operation will be performed.

* [CHORES] merge branch PBI-17 with staging and resolve migrations issue

Co-authored-by: johanes.christian <johanes.christian@ui.ac.id>
Co-authored-by: Johanes Christian <70894988+johaneschristian@users.noreply.github.com>